requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
click>=8.1.0
pydantic>=2.0.0
httpx>=0.24.0
tenacity>=8.2.0
python-dateutil>=2.8.0
pytz>=2023.3